refactor(core): better naming for url utils
